Split rv layouts unto their own uix files.
