+++ kivy/factory_registers.py
@@ -194,6 +194,6 @@ r('RecycleDataViewBehavior', module='kivy.uix.recycleview.views')
 r('RecycleDataAdapter', module='kivy.uix.recycleview.views')
 r('RecycleDataModelBehavior', module='kivy.uix.recycleview.datamodel')
 r('RecycleDataModel', module='kivy.uix.recycleview.datamodel')
-r('RecycleLayout', module='kivy.uix.layout')
-r('RecycleGridLayout', module='kivy.uix.gridlayout')
-r('RecycleBoxLayout', module='kivy.uix.boxlayout')
+r('RecycleLayout', module='kivy.uix.recyclelayout')
+r('RecycleGridLayout', module='kivy.uix.recyclegridlayout')
+r('RecycleBoxLayout', module='kivy.uix.recycleboxlayout')
